refactor: 숫자 입력값 검증 로직 통합
- 중복되는 숫자 입력 검증 로직을 parseNumberInput 메서드로 통합
- inputLottoPurchaseAmount와 inputWinningBonusNumber에서 공통 검증 로직 사용

+++ src/main/java/lotto/domain/BonusNumber.java
@@ -2,24 +2,24 @@ package lotto.domain;
 
 public class BonusNumber {
 
-    private final int number;
+    private final long number;
 
-    public BonusNumber(int number) {
+    public BonusNumber(long number) {
         validateNumberInRange(number);
         this.number = number;
     }
 
-    private static void validateNumberInRange(int number) {
+    private static void validateNumberInRange(long number) {
         if (isOutOfRange(number)) {
             throw new IllegalArgumentException("[ERROR] 보너스 번호는 1부터 45 사이의 숫자여야 합니다.");
         }
     }
 
-    private static boolean isOutOfRange(int number) {
+    private static boolean isOutOfRange(long number) {
         return number < 1 || number > 45;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
